requirements.txt
alabaster @ file:///home/ktietz/src/ci/alabaster_1611921544520/work
anyio @ file:///C:/ci/anyio_1620153418380/work/dist
appdirs==1.4.4
argh==0.26.2
argon2-cffi @ file:///C:/ci/argon2-cffi_1613037959010/work
astroid==2.3.3
async-generator @ file:///home/ktietz/src/ci/async_generator_1611927993394/work
atomicwrites==1.4.0
attrs @ file:///tmp/build/80754af9/attrs_1620827162558/work
autopep8 @ file:///tmp/build/80754af9/autopep8_1615918855173/work
Babel @ file:///tmp/build/80754af9/babel_1620871417480/work
backcall==0.2.0
basketball-reference-scraper==1.0.26
bcrypt @ file:///C:/ci/bcrypt_1597936263757/work
beautifulsoup4==4.10.0
black==19.10b0
bleach @ file:///tmp/build/80754af9/bleach_1628110601003/work
brotlipy==0.7.0
cachetools==4.2.4
certifi==2021.10.8
cffi @ file:///C:/ci/cffi_1625831749120/work
chardet==3.0.4
charset-normalizer==2.0.6
click @ file:///tmp/build/80754af9/click_1621604852318/work
cloudpickle @ file:///tmp/build/80754af9/cloudpickle_1632508026186/work
colorama==0.4.3
cryptography @ file:///C:/ci/cryptography_1616769344312/work
cycler==0.10.0
debugpy @ file:///C:/ci/debugpy_1629222854708/work
decorator==5.0.9
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
Deprecated==1.2.13
diff-match-patch @ file:///Users/ktietz/demo/mc3/conda-bld/diff-match-patch_1630511840874/work
docutils @ file:///C:/ci/docutils_1620828260973/work
entrypoints==0.3
et-xmlfile==1.1.0
flake8 @ file:///tmp/build/80754af9/flake8_1615834841867/work
future==0.18.2
google-api-core==2.0.1
google-auth==2.2.1
google-auth-oauthlib==0.4.6
google-cloud-bigquery==2.28.0
google-cloud-bigquery-storage==2.9.0
google-cloud-core==2.0.0
google-crc32c==1.2.0
google-resumable-media==2.0.3
googleapis-common-protos==1.53.0
grpcio==1.41.0
idna @ file:///tmp/build/80754af9/idna_1622654382723/work
imagesize @ file:///home/ktietz/src/ci/imagesize_1611921604382/work
importlib-metadata @ file:///C:/ci/importlib-metadata_1631916848487/work
intervaltree @ file:///Users/ktietz/demo/mc3/conda-bld/intervaltree_1630511889664/work
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.18.0
Jinja2 @ file:///tmp/build/80754af9/jinja2_1624781299557/work
json5 @ file:///tmp/build/80754af9/json5_1624432770122/work
jsonschema @ file:///Users/ktietz/demo/mc3/conda-bld/jsonschema_1630511932244/work
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-server @ file:///C:/ci/jupyter_server_1616084298403/work
jupyterlab @ file:///tmp/build/80754af9/jupyterlab_1635799997693/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
jupyterlab-server @ file:///tmp/build/80754af9/jupyterlab_server_1633419203660/work
keyring @ file:///C:/ci/keyring_1629321676013/work
kiwisolver==1.3.2
lazy-object-proxy==1.4.3
libcst==0.3.21
lxml==4.6.3
MarkupSafe @ file:///C:/ci/markupsafe_1621528314575/work
matplotlib==3.4.3
matplotlib-inline==0.1.2
mccabe==0.6.1
mistune==0.8.4
mypy-extensions==0.4.3
nbclassic @ file:///tmp/build/80754af9/nbclassic_1616085367084/work
nbclient @ file:///tmp/build/80754af9/nbclient_1614364831625/work
nbconvert @ file:///C:/ci/nbconvert_1624479163777/work
nbformat @ file:///tmp/build/80754af9/nbformat_1617383369282/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1613680548246/work
notebook @ file:///C:/ci/notebook_1635411854770/work
numpy==1.21.2
numpydoc @ file:///tmp/build/80754af9/numpydoc_1605117425582/work
oauthlib==3.1.1
openpyxl==3.0.9
packaging @ file:///tmp/build/80754af9/packaging_1625611678980/work
pandas==1.3.3
pandas-gbq==0.15.0
pandocfilters @ file:///C:/ci/pandocfilters_1605102497129/work
paramiko @ file:///tmp/build/80754af9/paramiko_1598886428689/work
parso==0.8.2
pathspec==0.7.0
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare==0.7.5
Pillow==8.3.2
pluggy @ file:///C:/ci/pluggy_1615976358795/work
prometheus-client @ file:///tmp/build/80754af9/prometheus_client_1623189609245/work
prompt-toolkit==3.0.18
proto-plus==1.19.2
protobuf==3.18.0
psutil @ file:///C:/ci/psutil_1612298324802/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybaseball==2.2.1
pycodestyle @ file:///home/ktietz/src/ci_mi/pycodestyle_1612807597675/work
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pydata-google-auth==1.2.0
pydocstyle @ file:///tmp/build/80754af9/pydocstyle_1621600989141/work
pyflakes @ file:///home/ktietz/src/ci_ipy2/pyflakes_1612551159640/work
PyGithub==1.55
Pygments==2.9.0
PyJWT==2.1.0
pylint==2.4.4
pyls-black @ file:///tmp/build/80754af9/pyls-black_1607553132291/work
pyls-spyder @ file:///tmp/build/80754af9/pyls-spyder_1613849700860/work
PyNaCl @ file:///C:/ci/pynacl_1595000047588/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
pyrsistent @ file:///C:/ci/pyrsistent_1600141795814/work
PySocks @ file:///C:/ci/pysocks_1605287845585/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-jsonrpc-server @ file:///tmp/build/80754af9/python-jsonrpc-server_1600278539111/work
python-language-server @ file:///tmp/build/80754af9/python-language-server_1607972495879/work
pytz==2021.3
pywin32==228
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq @ file:///C:/ci/pyzmq_1628276144921/work
QDarkStyle==2.8.1
QtAwesome @ file:///tmp/build/80754af9/qtawesome_1615991616277/work
qtconsole @ file:///tmp/build/80754af9/qtconsole_1632739723211/work
QtPy @ file:///tmp/build/80754af9/qtpy_1629397026935/work
regex @ file:///C:/ci/regex_1629302316714/work
requests @ file:///tmp/build/80754af9/requests_1629994808627/work
requests-oauthlib==1.3.0
rope @ file:///tmp/build/80754af9/rope_1623703006312/work
rsa==4.7.2
Rtree @ file:///C:/ci/rtree_1618421009405/work
schedule==1.1.0
scipy==1.7.1
Send2Trash @ file:///tmp/build/80754af9/send2trash_1632406701022/work
sip==4.19.13
six==1.16.0
sniffio @ file:///C:/ci/sniffio_1614030707456/work
snowballstemmer @ file:///tmp/build/80754af9/snowballstemmer_1611258885636/work
sortedcontainers @ file:///tmp/build/80754af9/sortedcontainers_1623949099177/work
soupsieve==2.2.1
Sphinx @ file:///tmp/build/80754af9/sphinx_1632506123190/work
sphinxcontrib-applehelp @ file:///home/ktietz/src/ci/sphinxcontrib-applehelp_1611920841464/work
sphinxcontrib-devhelp @ file:///home/ktietz/src/ci/sphinxcontrib-devhelp_1611920923094/work
sphinxcontrib-htmlhelp @ file:///tmp/build/80754af9/sphinxcontrib-htmlhelp_1623945626792/work
sphinxcontrib-jsmath @ file:///home/ktietz/src/ci/sphinxcontrib-jsmath_1611920942228/work
sphinxcontrib-qthelp @ file:///home/ktietz/src/ci/sphinxcontrib-qthelp_1611921055322/work
sphinxcontrib-serializinghtml @ file:///tmp/build/80754af9/sphinxcontrib-serializinghtml_1624451540180/work
spyder @ file:///C:/ci/spyder_1616776239898/work
spyder-kernels @ file:///C:/ci/spyder-kernels_1614030842607/work
terminado==0.9.4
testpath @ file:///tmp/build/80754af9/testpath_1624638946665/work
textdistance @ file:///tmp/build/80754af9/textdistance_1612461398012/work
three-merge @ file:///tmp/build/80754af9/three-merge_1607553261110/work
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
tornado==6.1
tqdm==4.62.3
traitlets==5.0.5
typed-ast @ file:///C:/ci/typed-ast_1624953785070/work
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1631814937681/work
typing-inspect==0.7.1
ujson @ file:///C:/ci/ujson_1611241570789/work
urllib3==1.26.7
watchdog @ file:///C:/ci/watchdog_1612471251191/work
wcwidth==0.2.5
webencodings==0.5.1
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306167264/work
wincertstore==0.2
wrapt==1.11.2
yapf @ file:///tmp/build/80754af9/yapf_1615749224965/work
zipp @ file:///tmp/build/80754af9/zipp_1625570634446/work
